[BE] refactor: 카테고리 리스트 파일 분기
- 카테고리 리스트 엘리먼트와, 카테고리 리스트 래퍼를 분리
